requirements.txt
pillow
pyautogui
ctypes